batch: Move ownership unit operations
Batch ownership has core claim metadata plus a neighboring
batch.ownership_units module for display-backed unit construction.

The enum, unit model, filtering, validation, and rebuild operations still
sit in batch.ownership, so the filesystem does not match the unit boundary.

This commit moves ownership-unit types and operations into
batch.ownership_units and updates production callers to use that module.

Subsequent commits will move tests and the architecture guard to the same
owner.
